[Fix] Install promtool v2.55.0
Problem:
https://gitea.obmondo.com/EnableIT/opsmondo/actions/runs/1731/jobs/2
During promtool CI test in Opsmondo, the alert tests are failing.

Cause:
As of now, we using ubuntu noble as base image, via which we're installing
promtool v2.45.3. This is causing problems, since it doesn't support
the tests we have written in Opsmondo.

Solution:
We're manually installing the latest promtool in v2 series.
This will fix the promtool CI in Opsmondo.

Resolves: https://gitea.obmondo.com/EnableIT/opsmondo/issues/95
Suggested-by: Ashish Jaiswal <ashish@obmondo.com>
Signed-off-by: Sidharth Jawale <sidharth@obmondo.com>